requirements.txt
python-telegram-bot==22.6
deep-translator==1.11.4
python-dotenv==1.2.1
httpx==0.28.1